+++ docs/examples/external_views/classes.py
@@ -1,7 +1,5 @@
-import pathlib
 import random
 
-import yaml
 from django.http import HttpRequest, JsonResponse
 from django.urls import include
 from django.views import View
@@ -10,6 +8,7 @@ from dmr.openapi import build_schema, load_schema
 from dmr.openapi.objects import PathItem
 from dmr.openapi.views import OpenAPIJsonView
 from dmr.routing import Router, path
+from examples.external_views.read_openapi import read_openapi_yaml
 
 
 class NumberView(View):
@@ -18,29 +17,16 @@ class NumberView(View):
 
 
 # Now, load the schema:
-
-raw_schema = yaml.safe_load(
-    pathlib.Path('examples/external_views/openapi.yml').read_text(
-        encoding='utf8',
-    ),
-)
+raw_schema = read_openapi_yaml('openapi.yml')
 
 # Create a router and URL patterns:
-
 router = Router(
     'api/',
-    [],
+    urls=[],
     external_urls=[  # pass `(external_url, path_item_spec)` pairs:
         (
-            path(
-                'number/',
-                NumberView.as_view(),
-                name='number',
-            ),
-            load_schema(
-                raw_schema['paths']['/api/number'],
-                PathItem,
-            ),
+            path('number/', NumberView.as_view(), name='number'),
+            load_schema(raw_schema['paths']['/api/number'], PathItem),
         ),
     ],
 )