+++ CHANGELOG.md
@@ -38,6 +38,7 @@ of requirements for an API to count as public.
   was renamed to `extra_namespace`, #1222
 - Changed ``skip_validation`` parameter to be kw-only
   on ``OpenAPIView.as_view()`` and all its subclasses, #1229
+- Renamed `Endpoint.get_schema` method to `Endpoint.get_operation`
 
 ### Features
 
@@ -47,6 +48,7 @@ of requirements for an API to count as public.
   and all its existing subclasses, #1222
 - Added an ability to load external OpenAPI schemas
   into our typed dataclasses, #1222
+- Added an option to skip some controllers / endpoints from the OpenAPI spec
 
 ### Bugfixes